Min-Heap
implemented a min-heap in Python, and that’s no small feat (although it could efficiently track the smallest feat).

To recap: MinHeap tracks the minimum element as the element at index 1 within an internal Python list.

When adding elements, we use .heapify_up() to compare the new element with its parent, making swaps if it violates the heap property: children must be greater than their parents.

When removing the minimum element, we swap it with the last element in the list. Then we use .heapify_down() to compare the new root with its children, swapping with the smaller child if necessary.